requirements.txt
streamlit>=1.28.0
google-generativeai>=0.3.0
reportlab>=4.0.0
python-dotenv>=1.0.0
Pillow>=10.0.0
pandas>=2.0.0
PyPDF2>=3.0.0
python-docx>=1.0.0